generators/requirements.txt
# Requirements for template generators
PyYAML>=6.0
